Extract StorageBucketAPI
